rename typo in filename and enable tests on ci
