+++ @vates/generator-toolbox/src/throttle.mts
