+++ @vates/generator-toolbox/src/throttle.test.mts
@@ -1,4 +1,4 @@
-import { GeneratorThrottler } from './trottle.mts'
+import { GeneratorThrottler } from './throttle.mts'
 import { suite, test } from 'node:test'
 import assert from 'node:assert'
